[MAR-5118] outbound_email.deleted cannot be NULL
The deleted column was added to outbound_email in 7.10. The default value of 0 was applied to all existing rows in all databases except for MSSQL/SQLServer. Having a NULL deleted column causes all sorts of havoc because the application assumes that the system row can be loaded from the database at any time. This upgrade script will correct this issue by always setting deleted to 0 if the application gets into a bad state.